+++ web/modules/custom/hoeringsportal_citizen_proposal/modules/hoeringsportal_citizen_proposal_fixtures/src/Fixture/CitizenProposalFixture.php
@@ -8,8 +8,15 @@ use Drupal\content_fixtures\Fixture\FixtureGroupInterface;
 use Drupal\Core\Datetime\DrupalDateTime;
 use Drupal\hoeringsportal_base_fixtures\Fixture\MediaFixture;
 use Drupal\hoeringsportal_base_fixtures\Fixture\ParagraphFixture;
-use Drupal\hoeringsportal_base_fixtures\Helper\Helper;
+use Drupal\hoeringsportal_base_fixtures\Helper\Helper as BaseFixtureHelper;
+use Drupal\hoeringsportal_citizen_proposal\Form\ProposalFormBase;
+use Drupal\hoeringsportal_citizen_proposal\Helper\Helper;
+use Drupal\hoeringsportal_citizen_proposal\Helper\MailHelper;
+use Drupal\hoeringsportal_citizen_proposal_archiving\Helper\Helper as ArchiveHelper;
 use Drupal\node\Entity\Node;
+use Drupal\node\NodeInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\Yaml\Yaml;
 
 /**
  * Page fixture.
@@ -22,8 +29,16 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
    * Constructor.
    */
   public function __construct(
-    readonly private Helper $baseFixtureHelper
-  ) {}
+    readonly private BaseFixtureHelper $baseFixtureHelper,
+    readonly private Helper $helper,
+    EventDispatcherInterface $eventDispatcher,
+    MailHelper $mailHelper,
+    ArchiveHelper $archivingHelper
+  ) {
+    // Prevent sending notification emails.
+    $eventDispatcher->removeSubscriber($mailHelper);
+    $eventDispatcher->removeSubscriber($archivingHelper);
+  }
 
   /**
    * {@inheritdoc}
@@ -33,7 +48,7 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
     $entity = Node::create([
       'type' => 'citizen_proposal',
       'title' => 'Borgerforslag nummer 1',
-      'status' => 1,
+      'status' => NodeInterface::PUBLISHED,
       'field_author_uuid' => '1111',
       'field_author_name' => 'Anders And',
       'field_author_email' => 'anders.and@itkdev.dk',
@@ -42,20 +57,25 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
       'field_content_state' => 'upcoming',
       'field_proposal' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlShort.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
       'field_remarks' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlShort.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
     ]);
     $entity->save();
     $this->addReference('node:citizen_proposal:Proposal1', $entity);
 
+    // Add some support.
+    for ($i = 0; $i < 87; $i++) {
+      $this->helper->saveSupport(uniqid('', TRUE), $entity, ['user_name' => self::class]);
+    }
+
     $entity = Node::create([
       'type' => 'citizen_proposal',
       'title' => 'Borgerforslag nummer 2',
-      'status' => 1,
+      'status' => NodeInterface::PUBLISHED,
       'field_author_uuid' => '2222',
       'field_author_name' => 'Fedtmule',
       'field_author_email' => 'fedtmule@itkdev.dk',
@@ -64,20 +84,25 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
       'field_content_state' => 'finished',
       'field_proposal' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtml1.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
       'field_remarks' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlLong.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
     ]);
     $entity->save();
     $this->addReference('node:citizen_proposal:Proposal2', $entity);
 
+    // Add some support.
+    for ($i = 0; $i < 3; $i++) {
+      $this->helper->saveSupport(uniqid('', TRUE), $entity, ['user_name' => self::class]);
+    }
+
     $entity = Node::create([
       'type' => 'citizen_proposal',
       'title' => 'Borgerforslag nummer 3',
-      'status' => 1,
+      'status' => NodeInterface::NOT_PUBLISHED,
       'field_author_uuid' => '3333',
       'field_author_name' => 'Hexia De Trick',
       'field_author_email' => 'givmiglykkemønten@itkdev.dk',
@@ -86,15 +111,25 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
       'field_content_state' => 'active',
       'field_proposal' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlLong.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
       'field_remarks' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlLong.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
     ]);
     $entity->save();
     $this->addReference('node:citizen_proposal:Proposal3', $entity);
+
+    // Set admin values.
+    $data = Yaml::parseFile(__DIR__ . '/CitizenProposalFixture/citizen_proposal_admin_form_values.yaml');
+    if (isset($data['citizen_proposal_admin_form_values'])) {
+      $values = $data['citizen_proposal_admin_form_values'];
+      /** @var \Drupal\node\Entity\NodeInterface $node */
+      $node = $this->getReference('node:landing_page:Proposals');
+      $values['approve_goto_url'] = $node->toUrl(options: ['alias' => TRUE])->toString();
+      $this->helper->setAdminValues($values);
+    }
   }
 
   /**
@@ -104,6 +139,7 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
     return [
       MediaFixture::class,
       ParagraphFixture::class,
+      CitizenProposalLandingPageFixture::class,
     ];
   }