+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Form/ProposalFormApprove.php
@@ -69,13 +69,13 @@ final class ProposalFormApprove extends ProposalFormBase {
       '#type' => 'container',
     ];
 
-    $emailHiddenText = $defaltValues['email_display'] ? '' : '<small><strong>(Hidden)</strong></small>';
+    $notShownOnProposalMessage = '<small><strong>' . $this->t('(Not shown on proposal)') . '</strong></small>';
     $form['author']['email_wrapper']['approve_form_email'] = [
       '#prefix' => '<h5>' . $this->t('E-mail') . '</h5>',
       '#type' => 'processed_text',
       '#text' => $defaltValues['email'],
       '#format' => 'filtered_html',
-      '#suffix' => $emailHiddenText,
+      '#suffix' => $defaltValues['email_display'] ? '' : $notShownOnProposalMessage,
     ];
 
     $form['author']['phone_display_wrapper'] = [
@@ -87,7 +87,7 @@ final class ProposalFormApprove extends ProposalFormBase {
       '#type' => 'processed_text',
       '#text' => $defaltValues['phone'],
       '#format' => 'filtered_html',
-      '#suffix' => '<small><strong>(Hidden)</strong></small>',
+      '#suffix' => $notShownOnProposalMessage,
     ];
 
     $form['approve_form_title'] = [
@@ -109,6 +109,28 @@ final class ProposalFormApprove extends ProposalFormBase {
       '#type' => 'processed_text',
       '#text' => $defaltValues['remarks'],
       '#format' => 'filtered_html',
+    ];
+
+    $form['consent'] = [
+      '#type' => 'checkbox',
+      '#title' => $this
+        ->t('Personal data storage consent'),
+      '#default_value' => TRUE,
+      '#description' => $this->getAdminFormStateValue('consent_help'),
+      '#attributes' => [
+        'disabled' => 'disabled',
+      ],
+    ];
+
+    $form['allow_email'] = [
+      '#type' => 'checkbox',
+      '#title' => $this
+        ->t('Allow email'),
+      '#default_value' => $defaltValues['allow_email'],
+      '#description' => $this->getAdminFormStateValue('allow_email_help'),
+      '#attributes' => [
+        'disabled' => 'disabled',
+      ],
     ];
 
     $form['actions']['#type'] = 'actions';
@@ -144,11 +166,25 @@ final class ProposalFormApprove extends ProposalFormBase {
       return $this->abandonSubmission();
     }
 
-    $this->applyRedirect($this->getAdminFormStateValue('approve_goto_url', '/'), $formState);
-
     $this->messenger()->addStatus($this->getAdminFormStateValue('approve_submission_text', $this->t('Thank you for your submission.')));
     $entity->save();
+
     $this->helper->deleteDraftProposal();
+
+    // Handle survey.
+    try {
+      if ($webform = $this->loadSurvey()) {
+        $this->webformHelper->saveSurveyResponse($webform, $entity);
+      }
+    }
+    catch (\Exception) {
+    }
+
+    $formState->setRedirectUrl(
+      $this->deAuthenticateUser(
+        $this->getAdminFormStateValueUrl('approve_goto_url', '/')
+      )
+    );
   }
 
   /**
@@ -158,7 +194,9 @@ final class ProposalFormApprove extends ProposalFormBase {
     $this->messenger()->addStatus($this->t('Your submission has been cancelled.'));
     $this->helper->deleteDraftProposal();
 
-    $this->applyRedirect($this->getAdminFormStateValue('approve_goto_url', '/'), $formState);
+    $formState->setRedirectUrl(
+      $this->getAdminFormStateValueUrl('approve_goto_url', '/')
+    );
   }
 
   /**
@@ -169,46 +207,6 @@ final class ProposalFormApprove extends ProposalFormBase {
     $url = Url::fromRoute('hoeringsportal_citizen_proposal.citizen_proposal.proposal_add');
 
     return new RedirectResponse($url->toString());
-  }
-
-  /**
-   * Add redirect to form state.
-   *
-   * @param string $inputUrl
-   *   The url added from admin interface.
-   * @param \Drupal\Core\Form\FormStateInterface $formState
-   *   The state of the form.
-   *
-   * @return \Drupal\Core\Form\FormStateInterface
-   *   The modified form.
-   */
-  private function applyRedirect(string $inputUrl, FormStateInterface &$formState): FormStateInterface {
-    $parsedUrl = parse_url($inputUrl);
-
-    try {
-      if (empty($parsedUrl['path'])) {
-        $formState->setRedirect('<front>');
-
-        return $formState;
-      }
-      $url = URL::fromUserInput($parsedUrl['path']);
-    }
-    catch (\Exception) {
-      $formState->setRedirect('<front>');
-
-      return $formState;
-    }
-
-    if ($url->isExternal() || !$url->isRouted()) {
-      $formState->setRedirect('<front>');
-    }
-    else {
-      $routeName = $url->getRouteName();
-      $routeParameters = $url->getRouteParameters();
-      $formState->setRedirect($routeName, $routeParameters);
-    }
-
-    return $formState;
   }
 
 }