+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Form/ProposalFormSupport.php
@@ -95,12 +95,33 @@ final class ProposalFormSupport extends ProposalFormBase {
 
     $form['email'] = [
       '#type' => 'email',
-      '#required' => TRUE,
       '#title' => $this
         ->t('Email'),
       '#default_value' => $defaltValues['email'],
       '#description' => $this->getAdminFormStateValue('support_email_help'),
       '#description_display' => 'before',
+    ];
+
+    $form['allow_email'] = [
+      '#type' => 'checkbox',
+      '#title' => $this
+        ->t('Allow email'),
+      '#default_value' => $defaltValues['support_allow_email_help'] ?? FALSE,
+      '#description' => $this->getAdminFormStateValue('support_allow_email_help'),
+      '#states' => [
+        'visible' => [
+          ':input[name="email"]' => ['filled' => TRUE],
+        ],
+      ],
+    ];
+
+    $form['consent'] = [
+      '#type' => 'checkbox',
+      '#title' => $this
+        ->t('Personal data storage consent'),
+      '#required' => TRUE,
+      '#default_value' => FALSE,
+      '#description' => $this->getAdminFormStateValue('consent_help'),
     ];
 
     $form['actions']['#type'] = 'actions';
@@ -145,7 +166,7 @@ final class ProposalFormSupport extends ProposalFormBase {
         [
           'user_name' => $form_state->getValue('name'),
           'user_email' => $form_state->getValue('email'),
-          'created' => time(),
+          'allow_email' => $form_state->getValue('allow_email'),
         ],
       );
       $this->messenger()->addStatus($this->getAdminFormStateValue('support_submission_text', $this->t('Thank you for your support.')));
@@ -154,8 +175,11 @@ final class ProposalFormSupport extends ProposalFormBase {
       $this->messenger()->addError($this->t('Something went wrong. Your support was not registered.'));
     }
 
-    $form_state
-      ->setRedirect('entity.node.canonical', ['node' => $node->id()]);
+    $form_state->setRedirectUrl(
+      $this->deAuthenticateUser(
+        $this->getAdminFormStateValueUrl('support_goto_url', NULL, $node->toUrl())
+      )
+    );
   }
 
 }